refactor: share 패키지 생성 및 관련 엔티티 이동
