+++ src/main/java/umc/stockoneqback/share/controller/ShareApiController.java
@@ -1,0 +1,11 @@
+package umc.stockoneqback.share.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class ShareApiController {
+}
