+++ src/main/java/umc/stockoneqback/share/service/ShareService.java
@@ -1,0 +1,11 @@
+package umc.stockoneqback.share.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class ShareService {
+}
